trader controller

+++ src/org/cz/portfolio/persistence/PortfolioDaoImpl.java
@@ -4,12 +4,16 @@ import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 import java.util.List;
+import java.util.Map;
+import java.util.TreeMap;
 
 import org.apache.log4j.Logger;
 import org.cz.home.persistence.PersistenceRuntimeException;
-import org.cz.portfolio.Portfolio;
-import org.cz.portfolio.PortfolioEntry;
-import org.cz.portfolio.hs.PortfolioEntryHs;
+import org.cz.json.portfolio.Portfolio;
+import org.cz.json.portfolio.PortfolioEntry;
+import org.cz.json.portfolio.PortfolioEntryType;
+import org.cz.json.portfolio.PortfolioWatch;
+import org.cz.json.portfolio.hs.PortfolioEntryHs;
 import org.cz.user.BaseUser;
 import org.springframework.dao.DataAccessException;
 import org.springframework.jdbc.core.BeanPropertyRowMapper;
@@ -28,7 +32,7 @@ public class PortfolioDaoImpl extends NamedParameterJdbcDaoSupport implements Po
 	}
 
 	@Override
-	public void deletePortfolio(final Portfolio portfolio) throws PersistenceRuntimeException {
+	public void deletePortfolio(final Portfolio portfolio)  {
 		try
 		{
 			getJdbcTemplate().update("DELETE FROM portfolio WHERE id = "
@@ -37,7 +41,10 @@ public class PortfolioDaoImpl extends NamedParameterJdbcDaoSupport implements Po
 				    	  	ps.setLong(1, portfolio.getId());
 				      }
 				    });
-			
+			for (PortfolioWatch watch : portfolio.getWatchList().values())
+			{
+				deletePortfolioWatch(watch);
+			}
 		}
 		catch (DataAccessException e)
 		{
@@ -47,7 +54,7 @@ public class PortfolioDaoImpl extends NamedParameterJdbcDaoSupport implements Po
 	}
 
 	@Override
-	public void storePortfolio(final BaseUser baseUser, final Portfolio portfolio) throws PersistenceRuntimeException {
+	public void storePortfolio(final BaseUser baseUser, final Portfolio portfolio)  {
 		try
 		{
 			KeyHolder keyHolder = new GeneratedKeyHolder();
@@ -75,7 +82,7 @@ public class PortfolioDaoImpl extends NamedParameterJdbcDaoSupport implements Po
 		}	
 	}
 	@Override
-	public List<Portfolio> getPortfolios(final BaseUser baseUser) throws PersistenceRuntimeException {
+	public Map<String,Portfolio> getPortfolios(final BaseUser baseUser)  {
 		try
 		{
 			final String sql = "SELECT * FROM portfolio WHERE baseuserid=?";
@@ -84,11 +91,13 @@ public class PortfolioDaoImpl extends NamedParameterJdbcDaoSupport implements Po
 				          preparedStatement.setObject(1,baseUser.getId());
 				        }
 				      }, BeanPropertyRowMapper.newInstance(Portfolio.class));
+			Map<String,Portfolio> mp = new TreeMap<String,Portfolio>();
 			for (Portfolio port : ports)
 			{
-				getPortfolioEntries(port);
+				getPortfolioWatchList(port);
+				mp.put(port.getName(),port);
 			}
-			return ports;
+			return mp;
 		}
 		catch (DataAccessException e)
 		{
@@ -96,68 +105,27 @@ public class PortfolioDaoImpl extends NamedParameterJdbcDaoSupport implements Po
 			throw new PersistenceRuntimeException("Could not execute : " + e.getMessage());
 		}
 	}
-
-	private void getPortfolioEntries(final Portfolio portfolio) {
-		try
-		{
-			final String sql = "SELECT * FROM portfolioentry WHERE portfolioid=? ORDER BY securityticker";
-			List<PortfolioEntry> ports = getJdbcTemplate().query(sql,new PreparedStatementSetter() {
-				        public void setValues(PreparedStatement preparedStatement) throws SQLException {
-				          preparedStatement.setLong(1,portfolio.getId());
-				        }
-				      }, BeanPropertyRowMapper.newInstance(PortfolioEntry.class));
-			for (PortfolioEntry port : ports)
-			{
-				getPortfolioEntryHss(port);
-				portfolio.getEntries().put(port.getSecurityTicker(), port);
-			}
-		}
-		catch (DataAccessException e)
-		{
-			log.error("Could not execute : " + e.getMessage());
-			throw new PersistenceRuntimeException("Could not execute : " + e.getMessage());
-		}
-	}
-
-	private void getPortfolioEntryHss(final PortfolioEntry portEntry) {
-		try
-		{
-			final String sql = "SELECT * FROM portfolioentryhs WHERE portfolioentryid=?";
-			List<PortfolioEntryHs> hsList = getJdbcTemplate().query(sql,new PreparedStatementSetter() {
-				        public void setValues(PreparedStatement preparedStatement) throws SQLException {
-				          preparedStatement.setObject(1,portEntry.getId());
-				        }
-				      }, BeanPropertyRowMapper.newInstance(PortfolioEntryHs.class));
-			portEntry.setHsList(hsList);
-		}
-		catch (DataAccessException e)
-		{
-			log.error("Could not execute : " + e.getMessage());
-			throw new PersistenceRuntimeException("Could not execute : " + e.getMessage());
-		}
-	}
-
+	
 	@Override
-	public void storePortfolioEntry(final Portfolio portfolio,final PortfolioEntry entry) {
-
+	public void storePortfolioWatch(final Portfolio portfolio, final PortfolioWatch watch) {
 		try
 		{
 			KeyHolder keyHolder = new GeneratedKeyHolder();
-			final String sql = "INSERT INTO portfolioentry (portfolioid,securityticker) VALUES (?,?)";
+			final String sql = "INSERT INTO portfoliowatch (portfolioid,ticker) VALUES (?,?)";
 			getJdbcTemplate().update(
 			    new PreparedStatementCreator() {
 			        public PreparedStatement createPreparedStatement(Connection connection) throws SQLException {
 			            PreparedStatement ps =
 			                connection.prepareStatement(sql, new String[] {"id"});
 			            ps.setLong(1, portfolio.getId());
-						ps.setString(2, entry.getSecurity().getTicker());
+						ps.setString(2, watch.getTicker());
 			            return ps;
 			        }
 			    },
 			    keyHolder);
 			
 			final long id = keyHolder.getKey().longValue();
-			entry.setId(id);
+			watch.setId(id);
 		}
 		catch (DataAccessException e)
 		{
@@ -167,27 +135,136 @@ public class PortfolioDaoImpl extends NamedParameterJdbcDaoSupport implements Po
 	}
 
 	@Override
-	public void storePortfolioEntryHs(final PortfolioEntryHs phs,final PortfolioEntry entry) throws DataAccessException
+	public void deletePortfolioWatch(final PortfolioWatch watch){
+		try
+		{
+			getJdbcTemplate().update("DELETE FROM portfoliowatch WHERE id = ?"
+					  , new PreparedStatementSetter() {
+						public void setValues(PreparedStatement ps) throws SQLException {
+			    	  	ps.setLong(1, watch.getId());
+			      }
+			    });
+			for (PortfolioEntry pe : watch.getEntries())
+			{
+				deletePortfolioEntry(pe);
+			}
+		}
+		catch (DataAccessException e)
+		{
+			log.error("Could not execute : " + e.getMessage());
+			throw new PersistenceRuntimeException("Could not execute : " + e.getMessage());
+		}	
+	}
+	
+	private void getPortfolioWatchList(final Portfolio portfolio) {
+		try
+		{
+			final String sql = "SELECT * FROM portfoliowatch WHERE portfolioid=?";
+			List<PortfolioWatch> watches = getJdbcTemplate().query(sql,new PreparedStatementSetter() {
+				        public void setValues(PreparedStatement preparedStatement) throws SQLException {
+				          preparedStatement.setLong(1,portfolio.getId());
+				        }
+				      }, BeanPropertyRowMapper.newInstance(PortfolioWatch.class));
+			for (PortfolioWatch watch : watches)
+			{
+				getPortfolioEntries(watch);
+				portfolio.getWatchList().put(watch.getTicker(), watch);
+			}
+		}
+		catch (DataAccessException e)
+		{
+			log.error("Could not execute : " + e.getMessage());
+			throw new PersistenceRuntimeException("Could not execute : " + e.getMessage());
+		}
+	}
+	
+	
+	private void getPortfolioEntries(final PortfolioWatch watch) {
+		try
+		{
+			final String sql = "SELECT * FROM portfolioentry WHERE portfoliowatchid=?";
+			List<PortfolioEntry> ports = getJdbcTemplate().query(sql,new PreparedStatementSetter() {
+				        public void setValues(PreparedStatement preparedStatement) throws SQLException {
+				          preparedStatement.setLong(1,watch.getId());
+				        }
+				      }, BeanPropertyRowMapper.newInstance(PortfolioEntry.class));
+			for (PortfolioEntry port : ports)
+			{
+				if (port.getType().equals(PortfolioEntryType.HockeyStick))
+					port = getPortfolioEntryHss(port);
+				watch.getEntries().add(port);
+			}
+		}
+		catch (DataAccessException e)
+		{
+			log.error("Could not execute : " + e.getMessage());
+			throw new PersistenceRuntimeException("Could not execute : " + e.getMessage());
+		}
+	}
+
+	private PortfolioEntry getPortfolioEntryHss(final PortfolioEntry portEntry) {
+		try
+		{
+			final String sql = "SELECT * FROM portfolioentryhs WHERE portfolioentryid=?";
+			PortfolioEntryHs portHs = getJdbcTemplate().queryForObject(sql,new Long[] { portEntry.getId() },BeanPropertyRowMapper.newInstance(PortfolioEntryHs.class));
+			portHs.setSecurityTicker(portEntry.getSecurityTicker());
+			portHs.setId(portEntry.getId());
+			portHs.setType(portEntry.getType());
+			return portHs;
+		}
+		catch (DataAccessException e)
+		{
+			log.error("Could not execute : " + e.getMessage());
+			throw new PersistenceRuntimeException("Could not execute : " + e.getMessage());
+		}
+	}
+
+	@Override
+	public void storePortfolioEntry(final PortfolioWatch watch,final PortfolioEntry entry) {
+
+		try
+		{
+			KeyHolder keyHolder = new GeneratedKeyHolder();
+			final String sql = "INSERT INTO portfolioentry (portfoliowatchid,securityticker) VALUES (?,?)";
+			getJdbcTemplate().update(
+			    new PreparedStatementCreator() {
+			        public PreparedStatement createPreparedStatement(Connection connection) throws SQLException {
+			            PreparedStatement ps =
+			                connection.prepareStatement(sql, new String[] {"id"});
+			            ps.setLong(1, watch.getId());
+						ps.setString(2, entry.getSecurityTicker());
+			            return ps;
+			        }
+			    },
+			    keyHolder);
+			
+			final long id = keyHolder.getKey().longValue();
+			entry.setId(id);
+			
+			if (entry.getType().equals(PortfolioEntryType.HockeyStick))
+				storePortfolioEntryHs((PortfolioEntryHs) entry);
+		}
+		catch (DataAccessException e)
+		{
+			log.error("Could not execute : " + e.getMessage());
+			throw new PersistenceRuntimeException("Could not execute : " + e.getMessage());
+		}	
+	}
+
+	private void storePortfolioEntryHs(final PortfolioEntryHs phs) throws DataAccessException
 	{
-		KeyHolder keyHolder = new GeneratedKeyHolder();
 		final String sql = "INSERT INTO portfolioentryhs (portfolioentryid,status,daycount,ceiling) VALUES (?,?,?,?)";
 		try
 		{
-			getJdbcTemplate().update(
-				    new PreparedStatementCreator() {
-				        public PreparedStatement createPreparedStatement(Connection connection) throws SQLException {
-				            PreparedStatement ps =
-				                connection.prepareStatement(sql, new String[] {"id"});
-				        	ps.setLong(1, entry.getId());
+			getJdbcTemplate().update(sql
+			        , new PreparedStatementSetter() {
+			      public void setValues(PreparedStatement ps) throws SQLException {
+				        	ps.setLong(1, phs.getId());
 							ps.setString(2, phs.getStatus().name());
 							ps.setInt(3, phs.getDayCount());
 							ps.setDouble(4, phs.getCeiling());
-				            return ps;
-				        }
-				    },
-				    keyHolder);
-			final long id = keyHolder.getKey().longValue();
-			phs.setId(id);
+			      }
+		    });
 		}
 		catch (DataAccessException e)
 		{
@@ -221,36 +298,17 @@ public class PortfolioDaoImpl extends NamedParameterJdbcDaoSupport implements Po
 	}
 
 	@Override
-	public void deletePortfolioEntryHs(final PortfolioEntryHs phs) throws DataAccessException {
-		
-		final String sql = "UPDATE portfolioentryhs SET status=?,daycount=?,ceiling=? WHERE id = ?";
-		
-		try
-		{
-			getJdbcTemplate().update("DELETE FROM portfolioentryhs WHERE id = ?"
-					  , new PreparedStatementSetter() {
-						public void setValues(PreparedStatement ps) throws SQLException {
-			    	  	ps.setLong(1, phs.getId());
-			      }
-			    });
-		}
-		catch (DataAccessException e)
-		{
-			log.error("Could not execute : " + e.getMessage());
-			throw new PersistenceRuntimeException("Could not execute : " + e.getMessage());
-		}	
-	}
-
-	@Override
 	public void deletePortfolioEntry(final PortfolioEntry entry) {
 		try
 		{
-			getJdbcTemplate().update("DELETE FROM portfolioentryhs WHERE portfolioentryid = ?"
-					  , new PreparedStatementSetter() {
-						public void setValues(PreparedStatement ps) throws SQLException {
-			    	  	ps.setLong(1, entry.getId());
-			      }
-			    });
+			if (entry.getType().equals(PortfolioEntryType.HockeyStick))
+				getJdbcTemplate().update("DELETE FROM portfolioentryhs WHERE portfolioentryid = ?"
+						  , new PreparedStatementSetter() {
+							public void setValues(PreparedStatement ps) throws SQLException {
+				    	  	ps.setLong(1, entry.getId());
+				      }
+				    });
+			
 			getJdbcTemplate().update("DELETE FROM portfolioentry WHERE id = ?"
 					  , new PreparedStatementSetter() {
 						public void setValues(PreparedStatement ps) throws SQLException {
@@ -264,5 +322,6 @@ public class PortfolioDaoImpl extends NamedParameterJdbcDaoSupport implements Po
 			throw new PersistenceRuntimeException("Could not execute : " + e.getMessage());
 		}	
 	}
-	
+
+
 }
